read repairing client for memcached
(sapling split of a7693d32d67b0798b7dd137001e63bb44484b10a)

+++ birdcage/finagle/finagle-redis/src/main/scala/com/twitter/finagle/redis/Client.scala
@@ -191,7 +191,7 @@ class Client(service: Service[Command, Reply]) {
    * @return Number of elements added to sorted set
    */
   def zAdd(key: Array[Byte], score: Double, member: Array[Byte]): Future[Int] =
-    doRequest(ZAdd(key, ZMember(score, member))) {
+    doRequest(ZAdd(key, ZMember(score.toFloat, member))) {
       case IntegerReply(n)     => Future.value(n)
     }
 
@@ -212,7 +212,7 @@ class Client(service: Service[Command, Reply]) {
    * @return Number of elements between min and max in sorted set
    */
   def zCount(key: Array[Byte], min: Double, max: Double): Future[Int] =
-    doRequest(ZCount(key, ZInterval(min), ZInterval(max))) {
+    doRequest(ZCount(key, ZInterval(min.toFloat), ZInterval(max.toFloat))) {
       case IntegerReply(n)     => Future.value(n)
     }
 
@@ -228,8 +228,8 @@ class Client(service: Service[Command, Reply]) {
     doRequest(
       ZRangeByScore(
         BytesToString(key),
-        ZInterval(min),
-        ZInterval(max),
+        ZInterval(min.toFloat),
+        ZInterval(max.toFloat),
         Some(WithScores),
         Some(Limit(offset, count))
       )
